Changes to make drawing much faster
Draw methods that are only needed when the graphics are displayed have been moved to the paint method of the content pane.
This makes it so that a load of slow drawing methods like fillRect can be avoided in every loop.
Dump stats are instead stored in a string now in ZooGas. It is only updated each time gameLoop is called now; this is a hacky way to avoid concurrency issues with iterating through particles while drawing.

Signed-off-by: asw12 <asw12@berkeley.edu>